+++ apps/community/Contacts/Components/FormContact.tsx
@@ -124,18 +124,19 @@ export default class FormContact<P, S> extends HubletoForm<FormContactProps,Form
     return <>
       <div className='card'>
         <div className='card-body flex flex-row gap-2'>
-          <div className="w-1/2">
+          <div className="w-3/4">
             <div className="flex gap-2 w-full">
               <div>
                 <i className="fas fa-user text-2xl p-4 text-gray-500"></i>
               </div>
               <div className="w-full">
                 {this.inputWrapper('salutation')}
-
                 <div className="flex gap-2">
-                  <div className="flex-1">{this.inputWrapper('first_name', {cssClass: 'text-2xl text-primary'})}</div>
-                  <div className="flex-1">{this.inputWrapper('middle_name', {cssClass: 'text-2xl text-primary'})}</div>
-                  <div className="flex-1">{this.inputWrapper('last_name', {cssClass: 'text-2xl text-primary'})}</div>
+                  <div className="flex-1">{this.inputWrapper('title_before')}</div>
+                  <div className="flex-2">{this.inputWrapper('first_name', {cssClass: 'text-2xl text-primary'})}</div>
+                  <div className="flex-2">{this.inputWrapper('middle_name', {cssClass: 'text-2xl text-primary'})}</div>
+                  <div className="flex-2">{this.inputWrapper('last_name', {cssClass: 'text-2xl text-primary'})}</div>
+                  <div className="flex-1">{this.inputWrapper('title_after')}</div>
                 </div>
               </div>
             </div>