+++ apps/community/Settings/Models/User.php
@@ -25,8 +25,8 @@ class User extends \ADIOS\Models\User
   {
     return array_merge(parent::describeColumns(), [
       'first_name' => (new Varchar($this, $this->translate('First name'))),
-      'middle_name' => (new Varchar($this, $this->translate('Middle name'))),
       'last_name' => (new Varchar($this, $this->translate('Last name'))),
+      'nick' => (new Varchar($this, $this->translate('Nick'))),
       'email' => (new Varchar($this, $this->translate('Email'))),
       'language' => (new Varchar($this, $this->translate('Language')))->setEnumValues(self::ENUM_LANGUAGES),
       'id_active_profile' => (new Lookup($this, $this->translate("Active profile"), Profile::class)),
@@ -62,11 +62,24 @@ class User extends \ADIOS\Models\User
   {
     $description = parent::describeTable();
 
-    $description->ui['title'] = 'Users';
+    $description->ui['title'] = '';
     $description->ui['addButtonText'] = 'Add User';
     $description->ui['showHeader'] = true;
     $description->ui['showFulltextSearch'] = true;
     $description->ui['showFooter'] = false;
+
+    $description->permissions['canDelete'] = false;
+
+    $description->columns = [
+      'first_name' => $description->columns['first_name'],
+      'last_name' => $description->columns['last_name'],
+      'nick' => $description->columns['nick'],
+      'email' => $description->columns['email'],
+      'language' => $description->columns['language'],
+      'id_active_profile' => $description->columns['id_active_profile'],
+      'is_active' => $description->columns['is_active'],
+      'roles' => (new Varchar($this, $this->translate('Roles'))),
+    ];
 
     return $description;
   }