+++ apps/community/Usage/Models/Log.php
@@ -32,6 +32,8 @@ class Log extends \HubletoMain\Core\Models\Model
     $description->permissions['canCreate'] = false;
     $description->permissions['canUpdate'] = false;
     $description->permissions['canDelete'] = false;
+    $description->ui['showFulltextSearch'] = true;
+    $description->ui['showColumnSearch'] = true;
     return $description;
   }