+++ src/core/AppManager.php
@@ -79,7 +79,7 @@ class AppManager
 
       foreach (scandir($communityRepoFolder) as $appFolder) {
         $manifestFile = $communityRepoFolder . '/' . $appFolder . '/manifest.yaml';
-        if (is_file($manifestFile)) {
+        if (@is_file($manifestFile)) {
           $manifest = (array) \Symfony\Component\Yaml\Yaml::parse(file_get_contents($manifestFile));
           $manifest['appType'] = \HubletoMain\Core\App::APP_TYPE_COMMUNITY;
           $appNamespaces['HubletoApp\\Community\\' . $appFolder] = $manifest;