+++ CHANGELOG.md
@@ -10,8 +10,10 @@
   * improved forms for deals and leads
   * various bugfixes and other improvements
   * added some translations
-  * added probability to the Deal form and Pipeline view and added weighted price
+  * added probability to the Deal form and Pipeline view and added weighted profit
   * massive rebranding and UI updates
+  * new sidebar configuration feature in settings panel
+  * improved default pipeline
 
 ## Release v0.10